Refactoring of Content component.
