--- src/Component/Content.css
@@ -1,77 +1,0 @@
-.Content {
- position: relative;
- top: 0px;
- left: 0px;
-    display: flex;
-    width: 100%;
-    height: 100%;
-    z-index: 1;
-}
-.blok_left {
-    width: 195px;
-    max-height: 100%;
-    background-color: rgb(84, 69, 69);
-    z-index: 2;
-}
-.Logo_Rolex{
-position: relative;
-top: 105px;
-left: 0px;
-width: 100%;
-padding: 10px; 
-cursor: pointer;
-}
-.Logo_PatekPhilippe {
-    position: relative;
-    top: 650px;
-    left: 0px;
-    width: 100%;
-    padding: 10px; 
-    cursor: pointer;
-}
-.blok_centre {
-    padding-top: 20px;
-    width: 100%;
-    max-height: 100%;
-    z-index: 1;
-}
-.blok_centre h1 {
-text-align: center;
-margin: 20px 0px 0px 0px;
-}
-.text,
-.text2 {
-    font-family: sans-serif;
-    font-size: 20px;
-    padding: 10px;
-    overflow: hidden;
-    width: 100%;
-    height: 700px; 
-}
-.blok_centre img {
-    padding: 6px;
-    float: left;
-    max-width: 50%;
-    max-height: 100%;
-}
-.Rolex_a a {
-    position: relative;
-    top: 50px;
-    left: 20px;
-    text-decoration: none;
-    color:rgb(84, 69, 69);
-}
-.Patek_a a {
-     position: relative;
-     top: 50px;
-    left: 20px;
-    text-decoration: none;
-    color:rgb(84, 69, 69);
-}
-.blok_right {
-    width: 195px;
-    max-height: 100%;
-    background-color: rgb(84, 69, 69);
-    z-index: 2;
-}
-
