--- src/Component/Content.js
@@ -1,90 +1,0 @@
-import React from "react";
-import './Content.css';
-import AllItemSlider from './AllItemSlider';
-
-export default function Content() {
-  return(
-    <div className="Content">
-      <div className="blok_left">
-        <a href="about"><img className="Logo_Rolex" src="https://freepngimg.com/download/rolex/28107-5-rolex-logo-photos.png" alt="#" /></a>
-        <a href="aboutPP"><img className="Logo_PatekPhilippe" src="https://cdn.worldvectorlogo.com/logos/patek-philippe-sa-1.svg" alt="#" /></a>
-      </div>
-
-      <div className="blok_centre">
-        <h1>Skeleton watch</h1>
-        <div className="text">
-          <div className="Rolex_a">
-            <a href="about"><span>Rolex</span></a>
-          </div>
-          
-          <img className="img_watch" src="/pictures/WatchFoto/2.0.jpg"  alt="#" />
-          <span>Skeleton</span> - the brightest and most unusual watch model,
-          equipped with transparent durable glass, through which neatly carved,
-          high-quality details are visible <br/>
-          <span>Skeleton</span> - this is the choice of aesthetes.
-          This watch is a nice and stylish accessory,
-          they amaze with originality.<br/>
-
-          The watch case is made of high quality stainless steel,
-          which guarantees the case’s color to remain intact for years.
-          Handmade leather bracelet. The strap keeps
-          the watch tightly on the wrist. Quality packaging comes
-          with the watch.
-          The company was founded in 1905 by two Germans,
-          Hans Wilsdorf and his son-in-law Alfred Davis.
-          Already in 1914, the British astronomical community
-          issued watches from Rolex an accuracy class of "A",
-          which only high-precision marine chronographs had.
-          In 1933, British explorers flying over Everest
-          had a good old Oyster on their hand. In 1956,
-          the Rolex Day-Date was introduced,
-          the world's first watch to display the full
-          name of the day of the week.
-          Rolex is a partner and sponsor
-          of many sporting events. The company
-          has been the official sponsor of the 24 Hours
-          of Le Mans since 2001. Rolex also sponsors
-          equestrian sports, is a partner in tennis
-          tournaments and golf competitions. That is,
-          the brand relies on sponsorship of the
-          aristocratic and elite sports listed above.
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Doloremque sunt molestiae voluptas, necessitatibus aperiam modi totam illo voluptatem rem vero dolore quos aliquid aut? Cum incidunt porro expedita quibusdam magnam.
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Excepturi dicta, doloribus minus odio ea eveniet fugit perspiciatis optio, aspernatur unde nisi magnam quidem eius quae error! Amet, error? Dolorum, aliquid?
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Repellendus ullam voluptates tenetur quas, repellat adipisci velit quia, necessitatibus accusamus architecto rerum voluptas neque rem reiciendis dolorum aliquam quos odio quis!
-
-        </div>
-        <div className="text2">
-          <div className="Patek_a">
-            <a href="aboutPP"><span>Patek Philippe</span></a>
-          </div>
-          <img className="img_watch2" src="https://media.gqmagazine.fr/photos/62694b8f4572c80413a97813/master/w_1600%2Cc_limit/nev1.jpg" alt="#"/>
-          <span>
-                        Patek Philippe SA is a Swiss watch manufacturer.
-                        Founded by two Polish emigrants. This brand is one
-                        of the most expensive watches. The symbol of the company is the cross of the Spanish Order of Calatrava. Under this name, the "initial" line of models of the company (from ~ $ 9000) is produced. On March 1, 1839, in Geneva, two emigrants from Poland, Antoine Norbert Patek de Pravzic (Polish cavalry officer) and Francis (Francois) Czapek (watchmaker) opened a small watch company. At first, they did not set themselves the goal of conquering the secular nobility. At the beginning, they produced 2 hours a year strictly to order, and in 1851 - up to 200 hours. But over time, the founders decided to experiment with the appearance of the models. They began to process in detail, polish, tune and decorate the case with precious stones. Over time, exquisite watches received the "perfection" award at the London Exhibition in 1858.
-                        In 1845, Chapek left the company due to a conflict with Patek. The manufactory changed its name to Patek&Cie.
-                        On May 15, 1845, the free place was taken by Jean Adrien Philippe, the master of the temporary manufactory, the creator of the “keyless” winding mechanism.
-                        In 1851 he becomes a full partner of Patek. The company received the almost modern name of Patek Philippe & Co.
-                        In collaboration, Philippe was responsible for the development and production of new models, while Patek traveled the world and found influential customers. The company's clients already included Queen Victoria, Charlotte Bronte, Albert Einstein, Leo Tolstoy, Tchaikovsky Pyotr Ilyich and others.
-                        In 1868, Patek Philippe produced the world's first wristwatch, which is still kept in the company's museum. This gave impetus to the creation of new patents and innovations - chronograph, perpetual calendar, minute repeater and more.
-                        In 1932-1933, the company produced "in a watch competition" for the New York banker Henry Graves the most complex movement at that time.
-                        In 1999, Graves' watch was sold at Sotheby's for $11,002,500, the absolute highest ever paid for a watch.
-                        In 1901, when the titled founders passed away, the company was renamed Ancienne Manufacture d'Horlogerie Patek Philippe & Cie, SA. In 1932, the brothers Jean and Charles Stern acquired it. Since then, the current name of the company has been established - Patek Philippe SA. In 2007, the company is owned and run by the third and fourth generation of the Stern family, father and son Charles and Thierry. All production and the museum is located in Geneva. About 1300 people are employed.
-                        Patek Philippe's annual circulation is just over 40,000 watches and pocket watches, a quarter of which are women's quartz models. The total number of components of such a production volume reaches about 15 million. The lion's share of these "little things" are made at new sites, namely, 400 (!) types of wheels are machined here alone. For each, even the smallest small wheel, it takes from 40 to 6o (!!) operations, including the final manual polishing in the so-called Beauty Center (“Beauty Center”). A detailed and thorough approach to details says a lot about the quality of the final product and the grandiose price.
-                        Each finished movement is tested in 7 stages for a total of 18 days, and the same time is spent checking each assembled and adjusted watch. Successfully passed all tests, watches are marked with two letters PP - Patek Philippe Seal. This corporate seal of quality was introduced in 2009 and greatly upset many contenders for leadership in Haute Horlogerie, and dealt a tangible blow to the impeccable image of the Hallmark of Geneva.
-                        In the modern catalog there are 45 serial movements of 23 calibers. Over the entire history of the company's existence, 600,000 hours have been made.
-                  </span>
-        </div>
-        <AllItemSlider />
-        <div className="text_slay">
-          <p>
-            Lorem ipsum dolor, sit amet consectetur adipisicing elit. Consequuntur veritatis explicabo odit deserunt, enim ducimus dignissimos libero neque voluptatum officia reiciendis voluptatibus, dolorum quod, repudiandae incidunt amet porro optio nostrum.
-            Perferendis quibusdam modi aut inventore doloribus voluptate quae eaque, facilis nesciunt provident a explicabo obcaecati consequatur ducimus quam magni ex in quas libero saepe, odit quisquam! Molestiae vel cumque placeat?
-            Quibusdam eum, ullam quae quasi voluptatum vel numquam voluptatem accusantium odit esse culpa incidunt fugiat similique, commodi praesentium ipsam. Nihil quod maxime aliquid voluptate laudantium sunt vero, sed natus temporibus.
-          </p>
-        </div>
-      </div>
-      <div className="blok_right" />
-    </div>
-  )
-}
